Dev/feat/filtering update

+++ beakpeek/lib/Model/bird.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_print
+
 class Bird {
   Bird({
     required this.pentad,
@@ -17,9 +19,15 @@ class Bird {
       commonSpecies: json['common_species'],
       genus: json['genus'],
       species: json['species'],
-      reportingRate: json['reportingRate'],
+      reportingRate: json['reportingRate'].toDouble(),
     );
   }
+
+  // @override
+  // String toString() {
+  //   return '$commonGroup $commonSpecies';
+  // }
+
   final String pentad;
   final int spp;
   final String commonGroup;
